# specs/api.openapi.yaml
openapi: 3.0.3
info:
  title: SmartDocConverter API
  description: |
    RESTful API for SmartDocConverter - a SaaS document conversion platform.

    ## Authentication
    - JWT tokens in httpOnly cookies (for web app)
    - Bearer tokens in Authorization header (for API access)

    ## Rate Limiting
    - Free tier: 5 conversions/day
    - Premium: Unlimited
    - Headers: X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset
  version: 1.0.0
  contact:
    name: SmartDocConverter Support
    email: support@smartdocconverter.com
servers:
  - url: https://smartdocconverter.com/api/v1
    description: Production
  - url: https://staging.smartdocconverter.com/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /subscriptions/checkout:
    post:
      tags: [Subscriptions]
      summary: Create Stripe checkout session
      operationId: createCheckoutSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
...
  /convert/compress-pdf:
    post:
      tags: [Conversions]
      summary: Compress PDF file
      operationId: compressPdf
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompressPdfRequest'
      responses:
        '202':
          description: Compression job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionJobResponse'
...
  /convert/split-pdf:
    post:
      tags: [Conversions]
      summary: Split PDF into multiple files
      operationId: splitPdf
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitPdfRequest'
      responses:
        '202':
          description: Split job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionJobResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, FILE_TOO_LARGE, INVALID_FILE_TYPE, PROCESSING_ERROR, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: object
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 2
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        emailVerified:
          type: boolean
        image:
          type: string
          format: uri
        tier:
          type: string
          enum: [free, premium]
        createdAt:
          type: string
          format: date-time
...
    CheckoutRequest:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [monthly, annual]
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
    CheckoutResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
        url:
          type: string
          format: uri
...
    ConversionJobResponse:
      type: object
      properties:
        success:
          type: boolean
        job:
          type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, processing, completed, failed]
            progress:
              type: integer
              minimum: 0
              maximum: 100
            type:
              type: string
            inputFile:
              type: string
            outputFile:
              type: string
              nullable: true
            downloadUrl:
              type: string
              format: uri
              nullable: true
            error:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
            completedAt:
              type: string
              format: date-time
              nullable: true
...
    CompressPdfRequest:
      type: object
      required: [fileId]
      properties:
        fileId:
          type: string
          format: uuid
        level:
          type: string
          enum: [low, medium, high]
          default: medium
...
    SplitPdfRequest:
      type: object
      required: [fileId, splitMode]
      properties:
        fileId:
          type: string
          format: uuid
        splitMode:
          type: string
          enum: [pages, ranges, every_n]
        pages:
          type: array
          items:
            type: integer
        ranges:
          type: array
          items:
            type: string
        everyN:
          type: integer
          minimum: 1
...